fix proform behavior after discard data; refactor cluster-pro util
